fix: Update all function signatures and structs from i64 to i32

+++ server/src/routes/admin.rs
@@ -8,7 +8,7 @@ use serde::Serialize;
 pub struct AdminInfo {
     pub message: String,
     pub telegram_id: i64,
-    pub user_id: i64,
+    pub user_id: i32,
 }
 
 #[derive(Serialize)]
